phoenix/requirements.txt
pylint
pytest
python-dotenv
psycopg2-binary
discord.py